requirements.txt
#
# This file is autogenerated by pip-compile
# To update, run:
#
#    pip-compile
#
aiohttp==3.7.4
    # via
    #   -r requirements.in
    #   elasticsearch
aioredis==1.3.1
    # via -r requirements.in
amqp==2.6.1
    # via kombu
asgiref==3.3.4
    # via django
async-timeout==3.0.1
    # via
    #   aiohttp
    #   aioredis
attrs==20.2.0
    # via
    #   aiohttp
    #   jsonschema
aws-requests-auth==0.4.3
    # via -r requirements.in
billiard==3.6.3.0
    # via celery
bleach==3.3.0
    # via -r requirements.in
boto3==1.17.44
    # via
    #   -r requirements.in
    #   tabulator
botocore==1.20.44
    # via
    #   boto3
    #   s3transfer
cached-property==1.5.2
    # via tableschema
cachetools==4.1.1
    # via zenpy
celery-redbeat==1.0.0
    # via -r requirements.in
celery==4.4.7
    # via
    #   -r requirements.in
    #   celery-redbeat
certifi==2020.6.20
    # via
    #   elastic-apm
    #   elasticsearch
    #   requests
    #   sentry-sdk
cffi==1.14.3
    # via cryptography
chardet==3.0.4
    # via
    #   aiohttp
    #   requests
    #   tabulator
click==7.1.2
    # via
    #   tableschema
    #   tabulator
cryptography==3.4.8
    # via -r requirements.in
dj-database-url==0.5.0
    # via -r requirements.in
django-admin-sortable2==1.0
    # via -r requirements.in
django-appconf==1.0.4
    # via django-compressor
django-better-admin-arrayfield==1.3.0
    # via -r requirements.in
django-ckeditor==5.9.0
    # via -r requirements.in
django-compressor==2.4
    # via -r requirements.in
django-csp==3.6
    # via -r requirements.in
django-db-geventpool==3.2.3
    # via -r requirements.in
django-dynamic-models-readonly==0.1.1
    # via -r requirements.in
django-environ==0.4.5
    # via -r requirements.in
django-extensions==2.2.9
    # via -r requirements.in
django-js-asset==1.2.2
    # via django-ckeditor
django-redis==4.12.1
    # via -r requirements.in
django-staff-sso-client==2.1.0
    # via -r requirements.in
django-waffle==2.0.0
    # via -r requirements.in
django==3.2.10
    # via
    #   -r requirements.in
    #   django-admin-sortable2
    #   django-appconf
    #   django-csp
    #   django-db-geventpool
    #   django-dynamic-models-readonly
    #   django-redis
    #   django-staff-sso-client
    #   djangorestframework
djangorestframework==3.11.2
    # via -r requirements.in
docopt==0.6.2
    # via notifications-python-client
ecs-logging==0.5.0
    # via -r requirements.in
elastic-apm==6.2.2
    # via -r requirements.in
elasticsearch[async]==7.11.0
    # via -r requirements.in
et-xmlfile==1.0.1
    # via openpyxl
future==0.18.2
    # via notifications-python-client
geoalchemy2==0.8.4
    # via -r requirements.in
gevent==20.6.2
    # via -r requirements.in
greenlet==0.4.16
    # via gevent
hawk-server-asyncio==0.0.13
    # via -r requirements.in
hiredis==1.1.0
    # via aioredis
idna==2.10
    # via
    #   requests
    #   yarl
ijson==3.1.3
    # via tabulator
isodate==0.6.0
    # via tableschema
jdcal==1.4.1
    # via openpyxl
jmespath==0.10.0
    # via
    #   boto3
    #   botocore
jsonlines==2.0.0
    # via tabulator
jsonschema==3.2.0
    # via tableschema
kombu==4.6.11
    # via celery
linear-tsv==1.1.0
    # via tabulator
markdown==3.2.2
    # via -r requirements.in
markupsafe==1.1.1
    # via -r requirements.in
mohawk==1.1.0
    # via -r requirements.in
monotonic==1.5
    # via notifications-python-client
multidict==4.7.6
    # via
    #   aiohttp
    #   yarl
notifications-python-client==5.7.0
    # via -r requirements.in
oauthlib==3.1.0
    # via requests-oauthlib
openpyxl==3.0.6
    # via tabulator
packaging==20.4
    # via bleach
paste==3.4.3
    # via -r requirements.in
psqlparse==1.0rc7
    # via -r requirements.in
psutil==5.7.2
    # via -r requirements.in
psycogreen==1.0.2
    # via
    #   -r requirements.in
    #   django-db-geventpool
psycopg2-binary==2.9.2
    # via -r requirements.in
pycparser==2.20
    # via cffi
pyjwt==1.7.1
    # via
    #   -r requirements.in
    #   notifications-python-client
pyparsing==2.4.7
    # via packaging
pyrsistent==0.17.3
    # via jsonschema
python-dateutil==2.8.1
    # via
    #   botocore
    #   celery-redbeat
    #   tableschema
    #   zenpy
python-dotenv==0.12.0
    # via -r requirements.in
pytz==2020.1
    # via
    #   celery
    #   django
    #   zenpy
rcssmin==1.0.6
    # via django-compressor
redis==3.5.3
    # via
    #   -r requirements.in
    #   celery-redbeat
    #   django-redis
requests-oauthlib==1.3.0
    # via django-staff-sso-client
requests==2.25.1
    # via
    #   -r requirements.in
    #   aws-requests-auth
    #   notifications-python-client
    #   requests-oauthlib
    #   tableschema
    #   tabulator
    #   zenpy
rfc3986==1.4.0
    # via tableschema
rjsmin==1.1.0
    # via django-compressor
s3transfer==0.3.3
    # via boto3
sentry-sdk==0.17.6
    # via -r requirements.in
six==1.15.0
    # via
    #   bleach
    #   django-compressor
    #   django-extensions
    #   isodate
    #   jsonschema
    #   linear-tsv
    #   mohawk
    #   packaging
    #   paste
    #   psqlparse
    #   python-dateutil
    #   tableschema
    #   tabulator
    #   tenacity
    #   zenpy
sqlalchemy==1.3.19
    # via
    #   -r requirements.in
    #   geoalchemy2
    #   tabulator
sqlparse==0.4.2
    # via django
tableschema==1.20.0
    # via -r requirements.in
tabulator==1.53.1
    # via tableschema
tenacity==6.2.0
    # via celery-redbeat
typing-extensions==4.0.0
    # via aiohttp
unicodecsv==0.14.1
    # via
    #   tableschema
    #   tabulator
urllib3==1.26.5
    # via
    #   botocore
    #   elastic-apm
    #   elasticsearch
    #   requests
    #   sentry-sdk
vine==1.3.0
    # via
    #   amqp
    #   celery
webencodings==0.5.1
    # via bleach
werkzeug==1.0.1
    # via -r requirements.in
xlrd==2.0.1
    # via tabulator
xlsxwriter==1.2.8
    # via -r requirements.in
yarl==1.5.1
    # via aiohttp
zenpy==2.0.21
    # via -r requirements.in
zope.event==4.5.0
    # via gevent
zope.interface==5.1.0
    # via gevent

# The following packages are considered to be unsafe in a requirements file:
# setuptools
